Address Copilot auth and user management comments

+++ api/src/database/migrations/0006_add_user_roles.sql
@@ -2,6 +2,6 @@ ALTER TABLE "users" ADD COLUMN "role" varchar(50) DEFAULT 'user' NOT NULL CHECK 
 --> statement-breakpoint
 ALTER TABLE "users" ADD COLUMN "session_version" integer DEFAULT 0 NOT NULL;
 --> statement-breakpoint
-CREATE TABLE IF NOT EXISTS "setup_locks" ("id" varchar(50) PRIMARY KEY NOT NULL, "locked_at" integer NOT NULL);
+CREATE TABLE IF NOT EXISTS "setup_locks" ("id" varchar(50) PRIMARY KEY NOT NULL, "locked_at" bigint NOT NULL);
 --> statement-breakpoint
-CREATE TABLE IF NOT EXISTS "admin_mutation_locks" ("id" varchar(50) PRIMARY KEY NOT NULL, "locked_at" integer NOT NULL);
+CREATE TABLE IF NOT EXISTS "admin_mutation_locks" ("id" varchar(50) PRIMARY KEY NOT NULL, "locked_at" bigint NOT NULL);
